@@ -1,19 +1,23 @@
 name: shelf_static
-version: 0.2.9+2
+version: 1.0.0-dev
 description: Static file server support for shelf
 repository: https://github.com/dart-lang/shelf_static
 
 environment:
-  sdk: '>=2.3.0 <3.0.0'
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dependencies:
-  convert: '>=1.0.0 <4.0.0'
-  http_parser: '>=0.0.2+2 <5.0.0'
-  mime: '>=0.9.0 <=1.0.0'
-  path: '>=1.1.0 <2.0.0'
-  shelf: '>=0.5.7 <2.0.0'
+  convert: ^3.0.0
+  http_parser: ^4.0.0
+  mime: ^1.0.0
+  path: ^1.8.0
+  shelf: ^1.0.0
 
 dev_dependencies:
-  args: '^1.0.0'
-  test: '^1.2.0'
-  test_descriptor: '^1.0.0'
+  args: ^2.0.0
+  test: ^1.16.0
+  test_descriptor: ^2.0.0
+
+dependency_overrides:
+  shelf_packages_handler: ^2.0.1
+  test: ^1.16.0
