@@ -11,15 +11,15 @@ final p.Context _ctx = p.url;
 
 /// Makes a simple GET request to [handler] and returns the result.
 Future<Response> makeRequest(Handler handler, String path,
-    {String handlerPath, Map<String, String> headers}) {
+    {String? handlerPath, Map<String, String>? headers}) {
   final rootedHandler = _rootHandler(handlerPath, handler);
   return Future.sync(() => rootedHandler(_fromPath(path, headers)));
 }
 
-Request _fromPath(String path, Map<String, String> headers) =>
+Request _fromPath(String path, Map<String, String>? headers) =>
     Request('GET', Uri.parse('http://localhost$path'), headers: headers);
 
-Handler _rootHandler(String path, Handler handler) {
+Handler _rootHandler(String? path, Handler handler) {
   if (path == null || path.isEmpty) {
     return handler;
   }
